bot-detection-demo/pyproject.toml
[project]
name = "bot-detection-demo"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "databricks-connect>=17.0.0,<=17.0.1",
    "dbldatagen>=0.4.0.post1",
    "xgboost>=2.0.0",
    "scikit-learn>=1.3.0",
    "mlflow>=2.14.0",
    "psycopg2-binary>=2.9.0",
    "requests>=2.31.0",
    "jmespath>=1.0.1",
    "pyparsing>=3.2.5",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.backends.legacy:build"
